Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/checked-expected.txt


PASS input@checked is immediately reflected to 'checked' IDL attribute 
